Refactor v3.2 and v3.1
