+++ content/messaging/agent-chat-api/changelog/index.mdx
@@ -38,7 +38,7 @@ This document is the record of all the changes in the **Agent Chat API**, Web an
 - The **Customer** ([Web](/messaging/agent-chat-api/v3.3/#customer) & [RTM](/messaging/agent-chat-api/v3.3/rtm-reference/#customer)) data structure:
   - has a new property for visits, `ended_at`.
   - has new properties: `followed`, `online`, `group_ids`, `state`.
-- The [**Form field**](/messaging/agent-chat-api/v3.3/#form-fields) event's `type` property has a new possible value, `subject`. It's a new type that can appear in ticket forms.
+- In the [**Filled form fields**](/messaging/agent-chat-api/v3.3/#filled-form-fields) and [**Form fields**](/messaging/agent-chat-api/v3.3/#form-fields) data structures, the `type` property has a new possible value, `subject`. It's a new type that can appear in ticket forms.
 - There's a new event type, [**Form**](/messaging/agent-chat-api/v3.3/#form) that allows for sending an empty form to fill out.
 
 ### Properties
@@ -137,7 +137,7 @@ This document is the record of all the changes in the **Agent Chat API**, Web an
   - has new fields: `previous_thread_id`, `next_thread_id`, `queue`, and `queues_duration`.
   - had the `order` and `timestamp` fields replaced with a new one, `created_at` (date & time in microseconds in UTC).
   - had the `restricted_access` field type changed to `string`. It contains the reason for access restriction.
-- The [**Form field**](/messaging/agent-chat-api/v3.2/#form-fields) data structure has a new parameter, `id` (`fields.answer.id` for the `group_chooser` field type). It's an identifier for each option Customers can choose.
+- The [**Filled form fields**](/messaging/agent-chat-api/v3.2/#filled-form-fields) data structure has a new parameter, `id` (`fields.answer.id` for the `group_chooser` field type). It's an identifier for each option Customers can choose.
 - The [**Chat summary**](/messaging/agent-chat-api/v3.2/#chat-summaries) data structure has a new field, `queue`, nested in `last_thread_summary`.
 
 ### Properties
