+++ content/messaging/agent-chat-api/rtm-reference/index.mdx
@@ -149,7 +149,7 @@ The **Filled form** event contains data from a form (prechat or postchat survey)
 | `recipients` | yes      | `string`  | Possible values: `all` (default), `agents`                               |
 | `properties` | no       | `object`  | [Properties](#properties)                                                |
 | `form_id`    | yes      | `string`  |                                                                          |
-| `fields`     | yes      | `array`   | The fields a form contains. See [form fields](#form-fields) for details. |
+| `fields`     | yes      | `array`   | The fields a form contains. See [filled form fields](#filled-form-fields) for details. |
 
 ### Response
 
@@ -163,7 +163,7 @@ The **Filled form** event contains data from a form (prechat or postchat survey)
 | `recipients` | always     |                                                                     |
 | `properties` | optionally |                                                                     |
 | `form_id`    | always     |                                                                     |
-| `fields`     | always     | An array of [form fields](#form-fields)                             |
+| `fields`     | always     | An array of [filled form fields](#filled-form-fields)                             |
 
 </Text>
 
@@ -401,7 +401,7 @@ Apart from [Events](#events) and [Users](#users), there are also other common da
 - [Access](#access)
 - [Chats](#chats)
 - [Chat summaries](#chat-summaries)
-- [Form fields](#form-fields)
+- [Filled form fields](#filled-form-fields)
 - [Properties](#properties)
 - [Threads](#threads)
 
@@ -428,14 +428,14 @@ Apart from [Events](#events) and [Users](#users), there are also other common da
 
 <CodeResponse version="v3.2" type="agent" title={'Sample Chat summary data structure'} json="chatSummary"/>
 
-## Form fields
+## Filled form fields
 
 A component of the [Filled form](#filled-form) event.
 
-<CodeResponse version="v3.2" type="agent" title={'Sample Form fields'} json="formFields"/>
+<CodeResponse version="v3.2" type="agent" title={'Sample Filled form fields'} json="filledFormFields"/>
 
 | Field             | Required | Data type          |                                                                                                    Notes |
-| ----------------- | -------- | ------------------ | -------------------------------------------------------------------------------------------------------: |
+| ----------------- | -------- | ------------------ | ------------------------------------------------------------------------------------------------------- |
 | `fields`          | yes      | `array of objects` |                                                                              The fields a form contains. |
 | `type`            | yes      | `string`           | Possible values: `checkbox`, `email`, `name`, `question`, `textarea`, `group_chooser`, `radio`, `select` |
 | `id`              | yes      | `string`           |                                                                            Field id, for all field types |
