+++ content/messaging/customer-chat-api/changelog/index.mdx
@@ -34,7 +34,7 @@ This document is the record of all the changes in the **Customer Chat API**, Web
 - The **Rich Message** ([Web](/messaging/customer-chat-api/v3.3/#rich-message) & [RTM](/messaging/customer-chat-api/v3.3/rtm-reference/#rich-message)) data structure:
   - has a new parameter for images, `alternative_text`,
   - has a new property for buttons, `target`.
-- The [**Form field**](/messaging/customer-chat-api/v3.3/#form-fields) event's `type` property has a new possible value, `subject`. It's a new type that can appear in ticket forms.
+- In the [**Filled form fields**](/messaging/customer-chat-api/v3.3/#filled-form-fields) and [**Form fields**](/messaging/customer-chat-api/v3.3/#form-fields) data structures, the `type` property has a new possible value, `subject`. It's a new type that can appear in ticket forms.
 - There's a new event type, [**Form**](/messaging/customer-chat-api/v3.3/#form) that allows for sending an empty form to fill out.
 
 ### Greetings
