+++ content/messaging/customer-chat-api/rtm-reference/index.mdx
@@ -145,7 +145,7 @@ The **Filled form** event contains data from a form (prechat or postchat survey)
 | `type`       | yes      | `string`  | `filled_form`                                                            |
 | `properties` | no       | `object`  | [Properties](#properties)                                                |
 | `form_id`    | yes      | `string`  |                                                                          |
-| `fields`     | yes      | `array`   | The fields a form contains. See [form fields](#form-fields) for details. |
+| `fields`     | yes      | `array`   | The fields a form contains. See [filled form fields](#filled-form-fields) for details. |
 
 ### Response
 
@@ -158,7 +158,7 @@ The **Filled form** event contains data from a form (prechat or postchat survey)
 | `author_id`  | always     |                                                                     |
 | `properties` | optionally |                                                                     |
 | `form_id`    | always     |                                                                     |
-| `fields`     | always     | An array of [form fields](#form-fields)                             |
+| `fields`     | always     | An array of [filled form fields](#filled-form-fields)                             |
 
 </Text>
 
@@ -377,7 +377,7 @@ Apart from [Events](#events) and [Users](#users), there are also other common da
 - [Access](#access)
 - [Chats](#chats)
 - [Chat summaries](#chat-summaries)
-- [Form fields](#form-fields)
+- [Filled form fields](#filled-form-fields)
 - [Properties](#properties)
 - [Threads](#threads)
 
@@ -404,14 +404,14 @@ Apart from [Events](#events) and [Users](#users), there are also other common da
 
 <CodeResponse version="v3.2" type="customer" title={'Sample Chat summary data structure'} json="chatSummary"/>
 
-## Form fields
+## Filled form fields
 
 A component of the [Filled form](#filled-form) event.
 
-<CodeResponse version="v3.2" type="customer" title={'Sample Form fields'} json="formFields"/>
+<CodeResponse version="v3.2" type="customer" title={'Sample Filled form fields'} json="filledFormFields"/>
 
 | Field             | Required | Data type          |                                                                                                    Notes |
-| ----------------- | -------- | ------------------ | -------------------------------------------------------------------------------------------------------: |
+| ----------------- | -------- | ------------------ | ------------------------------------------------------------------------------------------------------- |
 | `fields`          | yes      | `array of objects` |                                                                              The fields a form contains. |
 | `type`            | yes      | `string`           | Possible values: `checkbox`, `email`, `name`, `question`, `textarea`, `group_chooser`, `radio`, `select` |
 | `id`              | yes      | `string`           |                                                                            Field id, for all field types |
