+++ content/messaging/customer-chat-api/v3.1/rtm-reference/index.mdx
@@ -598,7 +598,7 @@ Apart from [Events](#events) and [Users](#users), there are also other common da
 - [Access](#access)
 - [Chats](#chats)
 - [Chat summaries](#chat-summaries)
-- [Form fields](#form-fields)
+- [Filled form fields](#filled-form-fields)
 - [Properties](#properties)
 - [Threads](#threads)
 
@@ -713,12 +713,12 @@ Apart from [Events](#events) and [Users](#users), there are also other common da
 
 </CodeResponse>
 
-## Form fields
+## Filled form fields
 
 A component of the [Filled form](#filled-form) event.
 
 | Field             | Required | Data type          |                                                                                                    Notes |
-| ----------------- | -------- | ------------------ | -------------------------------------------------------------------------------------------------------: |
+| ----------------- | -------- | ------------------ | ------------------------------------------------------------------------------------------------------- |
 | `fields`          | yes      | `array of objects` |                                                                              The fields a form contains. |
 | `type`            | yes      | `string`           | Possible values: `checkbox`, `email`, `name`, `question`, `textarea`, `group_chooser`, `radio`, `select` |
 | `id`              | yes      | `string`           |                                                                            Field id, for all field types |
