+++ payloads/messaging/v3.2/agent-chat-api/index.js
@@ -12,6 +12,6 @@ export { default as customer } from "./users/customer.json"
 export { default as access } from "./other-structures/access.json"
 export { default as chat } from "./other-structures/chat.json"
 export { default as chatSummary } from "./other-structures/chatSummary.json"
-export { default as formFields } from "./other-structures/formFields.json"
+export { default as filledFormFields } from "./other-structures/filledFormFields.json"
 export { default as properties } from "./other-structures/properties.json"
 export { default as thread } from "./other-structures/thread.json"
